# swagger-user.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ember API 서버",
...
    "version": "v2.2"
  },
  "servers": [
    {
      "url": "https://ember-app.duckdns.org",
      "description": "Ember 배포 서버"
    }
  ],
...
  "paths": {
...
    "/api/support/inquiry": {
      "post": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "1:1 문의 접수",
        "description": "1:1 문의를 접수합니다.\n\n**요청 필드:**\n- `category` (필수): 문의 카테고리 (예: \"기능 문의\", \"버그 제보\")\n- `title` (필수): 제목\n- `content` (필수): 문의 내용\n\n**제한:** 진행 중(PENDING) 문의 5건 초과 시 SP001 에러\n\n**에러:** SP001(5건 초과)",
        "operationId": "createInquiry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquiryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "inquiryId": 1,
                    "createdAt": "2026-04-30T10:00:00"
                  }
                }
              }
            }
          },
          "429": {
            "description": "문의 제한 초과",
            "content": {
              "application/json": {
                "example": {
                  "code": "SP001",
                  "message": "문의는 최대 5건까지 접수할 수 있습니다",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/support/inquiries": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "내 문의 목록 조회",
        "description": "내 문의 목록을 조회합니다.\n\n**응답:** inquiryId, category, title, status(PENDING/ANSWERED/CLOSED), createdAt",
        "operationId": "getMyInquiries",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": [
                    {
                      "inquiryId": 1,
                      "category": "기능 문의",
                      "title": "테스트",
                      "status": "PENDING"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/support/inquiries/{inquiryId}": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "문의 상세 조회",
        "description": "> 📱 **화면:** 14.3 FAQ 및 고객센터 — 문의 목록에서 카드 탭 (상세 진입)",
        "operationId": "getInquiryDetail",
        "parameters": [
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "inquiryId": 1,
                    "category": "기능 문의",
                    "title": "테스트 문의",
                    "content": "문의 내용...",
                    "answer": null,
                    "status": "PENDING"
                  }
                }
              }
            }
          },
          "404": {
            "description": "문의 없음",
            "content": {
              "application/json": {
                "example": {
                  "code": "SP003",
                  "message": "문의를 찾을 수 없습니다",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/notices": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "공지사항 목록 조회",
        "description": "공지사항 목록을 조회합니다.\n\n**정렬:** 고정(isPinned=true) 우선, 그 다음 최신순\n- Redis 1시간 캐싱 적용",
        "operationId": "getNotices",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": [
                    {
                      "id": 1,
                      "title": "서비스 업데이트",
                      "category": "GENERAL",
                      "priority": "NORMAL",
                      "isPinned": true,
                      "publishedAt": "2026-04-30T10:00:00"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/notices/{noticeId}": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "공지사항 상세 조회",
        "description": "> 📱 **화면:** 14.2 앱 공지사항 — 공지 카드 탭 (상세 진입)",
        "operationId": "getNoticeDetail",
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": {
                    "id": 1,
                    "title": "서비스 업데이트",
                    "content": "업데이트 내용...",
                    "category": "SERVICE",
                    "priority": "HIGH",
                    "isPinned": true,
                    "publishedAt": "2026-04-30T10:00:00",
                    "viewCount": 42
                  }
                }
              }
            }
          },
          "404": {
            "description": "공지 없음",
            "content": {
              "application/json": {
                "example": {
                  "code": "NT001",
                  "message": "공지사항을 찾을 수 없습니다",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/notices/unread-count": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "미읽음 공지사항 수 조회",
        "description": "미읽음 공지 수를 반환합니다.\n\n**동작:** lastReadNoticeId 기준으로 그 이후 공지 수 카운트",
        "operationId": "getUnreadCount",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": 5
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/notices/banners": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "활성 배너 조회 (최대 5개)",
        "description": "현재 활성화된 배너를 조회합니다. 최대 5개.\n\n**응답:** id, title, imageUrl, linkType(NONE/INTERNAL/EXTERNAL), linkUrl\n- Redis 1시간 캐싱",
        "operationId": "getActiveBanners",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": [
                    {
                      "id": 1,
                      "title": "이벤트",
                      "imageUrl": "https://...",
                      "linkType": "NONE"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/faq": {
      "get": {
        "tags": [
          "Notice/Support"
        ],
        "summary": "FAQ 목록 조회",
        "description": "FAQ 목록을 조회합니다.\n\n**응답:** id, question, answer, category\n- Redis 1시간 캐싱",
        "operationId": "getFaqs",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "example": {
                  "code": "200",
                  "message": "OK",
                  "data": [
                    {
                      "id": 1,
                      "category": "MATCHING",
                      "question": "교환일기는 어떻게 시작하나요?",
                      "answer": "일기를 작성하고..."
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InquiryRequest": {
        "type": "object",
        "description": "1:1 문의 접수 요청",
        "properties": {
          "category": {
            "type": "string",
            "description": "문의 카테고리",
            "example": "ACCOUNT"
          },
          "title": {
            "type": "string",
            "description": "문의 제목 (5~100자)",
            "maxLength": 100,
            "minLength": 5
          },
          "content": {
            "type": "string",
            "description": "문의 내용 (10~2000자)",
            "maxLength": 2000,
            "minLength": 10
          }
        },
        "required": [
          "category",
          "content",
          "title"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}